+++ h2integrate/converters/iron/martin_mine_perf_model.py
@@ -35,6 +35,7 @@ class MartinIronMinePerformanceComponent(PerformanceModelBaseClass):
         3600,
         3600,
     )  # (min, max) time step lengths (in seconds) compatible with this model
+    _control_classifier = "dispatchable"
 
     def initialize(self):
         super().initialize()
@@ -76,13 +77,13 @@ class MartinIronMinePerformanceComponent(PerformanceModelBaseClass):
             desc="Crude ore input",
         )
 
-        # Default the ore set point input as the rated capacity
+        # Default the ore command value input as the rated capacity
         self.add_input(
-            "iron_ore_set_point",
+            "iron_ore_command_value",
             val=self.config.max_ore_production_rate_tonnes_per_hr,
             shape=n_timesteps,
             units="t/h",
-            desc="Iron ore set point for iron mine",
+            desc="Iron ore command value for iron mine",
         )
 
         self.add_output(
@@ -202,11 +203,11 @@ class MartinIronMinePerformanceComponent(PerformanceModelBaseClass):
         max_crude_ore_consumption = inputs["system_capacity"] * crude_ore_usage_per_processed_ore
         max_energy_consumption = inputs["system_capacity"] * energy_usage_per_processed_ore
 
-        # iron ore set point, saturated at maximum rated system capacity
-        processed_ore_set_point = np.where(
-            inputs["iron_ore_set_point"] > inputs["system_capacity"],
+        # iron ore command value, saturated at maximum rated system capacity
+        processed_ore_command_value = np.where(
+            inputs["iron_ore_command_value"] > inputs["system_capacity"],
             inputs["system_capacity"],
-            inputs["iron_ore_set_point"],
+            inputs["iron_ore_command_value"],
         )
 
         # available feedstocks, saturated at maximum system feedstock consumption
@@ -226,9 +227,13 @@ class MartinIronMinePerformanceComponent(PerformanceModelBaseClass):
         processed_ore_from_electricity = energy_available / energy_usage_per_processed_ore
         processed_ore_from_crude_ore = crude_ore_available / crude_ore_usage_per_processed_ore
 
-        # output is minimum between available feedstocks and output set point
+        # output is minimum between available feedstocks and output command value
         processed_ore_production = np.minimum.reduce(
-            [processed_ore_from_crude_ore, processed_ore_from_electricity, processed_ore_set_point]
+            [
+                processed_ore_from_crude_ore,
+                processed_ore_from_electricity,
+                processed_ore_command_value,
+            ]
         )
 
         # energy consumption