+++ h2integrate/converters/methanol/methanol_baseclass.py
@@ -38,6 +38,7 @@ class MethanolPerformanceBaseClass(PerformanceModelBaseClass):
         3600,
         3600,
     )  # (min, max) time step lengths (in seconds) compatible with this model
+    _control_classifier = "dispatchable"
 
     def initialize(self):
         super().initialize()