+++ h2integrate/converters/nitrogen/simple_ASU.py
@@ -69,6 +69,7 @@ class SimpleASUPerformanceModel(PerformanceModelBaseClass):
         3600,
         3600,
     )  # (min, max) time step lengths (in seconds) compatible with this model
+    _control_classifier = "dispatchable"
 
     def initialize(self):
         super().initialize()