+++ h2integrate/converters/nuclear/nuclear_plant.py
@@ -37,6 +37,7 @@ class QuinnNuclearPerformanceModel(PerformanceModelBaseClass):
         3600,
         3600,
     )  # (min, max) time step lengths (in seconds) compatible with this model
+    _control_classifier = "fixed"
 
     def initialize(self):
         super().initialize()
@@ -60,18 +61,18 @@ class QuinnNuclearPerformanceModel(PerformanceModelBaseClass):
             desc="Nuclear plant rated capacity",
         )
         self.add_input(
-            f"{self.commodity}_set_point",
+            f"{self.commodity}_command_value",
             val=self.config.system_capacity_kw,
             shape=n_timesteps,
             units=self.commodity_rate_units,
-            desc="Electricity set point for nuclear plant",
+            desc="Electricity command value for nuclear plant",
         )
 
     def compute(self, inputs, outputs):
         system_capacity = inputs["system_capacity"]
-        electricity_set_point = inputs[f"{self.commodity}_set_point"]
+        electricity_command_value = inputs[f"{self.commodity}_command_value"]
 
-        electricity_out = np.minimum(electricity_set_point, system_capacity)
+        electricity_out = np.minimum(electricity_command_value, system_capacity)
         electricity_out = np.clip(electricity_out, 0.0, system_capacity)
 
         outputs["electricity_out"] = electricity_out