+++ h2integrate/core/test/test_framework.py
@@ -455,7 +455,7 @@ def test_technology_connections(temp_dir):
     h2i_model = H2IntegrateModel(temp_highlevel_yaml)
     demand_profile = np.ones(8760) * 720.0
     h2i_model.setup()
-    h2i_model.prob.set_val("battery.electricity_demand", demand_profile, units="MW")
+    h2i_model.prob.set_val("battery.electricity_set_point", demand_profile, units="MW")
     h2i_model.run()