+++ h2integrate/finances/profast_lco.py
@@ -1,3 +1,4 @@
+import warnings
 from pathlib import Path
 
 import numpy as np
@@ -106,6 +107,18 @@ class ProFastLCO(ProFastBase):
         """
         pf = self.populate_profast(inputs)
 
+        if "system_level_control" in self.options["plant_config"] and np.all(
+            inputs["capacity_factor"] == 0.0
+        ):
+            outputs[self.LCO_str] = 1e12
+            msg = (
+                f"Commodity stream for finance group has a zero capacity factor. "
+                "If you recieve this warning multiple times, there may be a problem "
+                "with your setup. ProFAST is not being run on this iteration and the "
+                f"{self.LCO_str} is being set to default value of 1e12 ({self.price_units})"
+            )
+            warnings.warn(msg, UserWarning)
+            return
         # simulate ProFAST
         sol, summary, price_breakdown = run_profast(pf)